requirements.txt
opentelemetry-api==1.34.1
opentelemetry-distro==0.55b1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-sdk==1.34.1
